refactor: move lifecycle tests to test_reporting_lifecycle.py and rename tag test functions
